+++ src/components/map/aircraft/MapAircraftList.vue
@@ -56,11 +56,12 @@ function receiveMessage(event: MessageEvent) {
     if (event.origin !== config.public.DOMAIN) {
         return;
     }
+
     if (event.source === window) return; // the message is from the same window, so we ignore it
 
     if (event.data && 'selectedPilot' in event.data) {
         if (event.data.selectedPilot === null) {
-            mapStore.overlays = mapStore.overlays.filter(x => x.type !== 'pilot' || x.key !== hoveredAircraft.value?.toString());
+            mapStore.overlays = mapStore.overlays.filter(x => x.type !== 'pilot');
         }
         else {
             mapStore.addPilotOverlay(event.data.selectedPilot.toString());