+++ src/components/map/filters/settings/MapSettingsColors.vue
@@ -6,8 +6,10 @@
             </div>
             <div class="colors__button">
                 <common-button
+                    v-if="!!store.mapSettings.colors?.[isLightTheme ? 'light' : 'default']"
                     size="S"
                     type="link"
+                    @click="themeSyncActive = true"
                 >
                     Sync themes
                 </common-button>
@@ -25,14 +27,28 @@
             :model-value="store.mapSettings.colors?.[themeKey]?.firs ?? reactive({ color: 'success500', transparency: 0.1 })"
             @update:modelValue="setUserMapSettings({ colors: { [themeKey]: { firs: $event } } })"
         >
-            FIR
+            FIR / ARTCC
         </common-color>
         <common-color
             :default-color="{ color: 'info400', transparency: 0.1 }"
             :model-value="store.mapSettings.colors?.[themeKey]?.uirs ?? reactive({ color: 'info400', transparency: 0.1 })"
             @update:modelValue="setUserMapSettings({ colors: { [themeKey]: { uirs: $event } } })"
         >
-            UIR
+            UIR / FSS
+        </common-color>
+        <common-color
+            :default-color="{ color: 'lightgray150' }"
+            :model-value="store.mapSettings.colors?.[themeKey]?.centerText ?? reactive({ color: 'lightgray150' })"
+            @update:modelValue="setUserMapSettings({ colors: { [themeKey]: { centerText: $event } } })"
+        >
+            Center label (text)
+        </common-color>
+        <common-color
+            :default-color="{ color: 'darkgray850' }"
+            :model-value="store.mapSettings.colors?.[themeKey]?.centerBg ?? reactive({ color: 'darkgray850' })"
+            @update:modelValue="setUserMapSettings({ colors: { [themeKey]: { centerBg: $event } } })"
+        >
+            Center label (background)
         </common-color>
         <common-color
             :default-color="{ color: 'error300', transparency: 0.7 }"
@@ -40,6 +56,30 @@
             @update:modelValue="setUserMapSettings({ colors: { [themeKey]: { runways: $event } } })"
         >
             Runways
+        </common-color>
+        <common-color
+            :default-color="{ transparency: 1 }"
+            :model-value="store.mapSettings.colors?.[themeKey]?.gates ? { transparency: store.mapSettings.colors?.[themeKey]?.gates ?? 1 } : reactive({ transparency: 1 })"
+            transparency-only
+            @update:modelValue="setUserMapSettings({ colors: { [themeKey]: { gates: $event.transparency } } })"
+        >
+            Gates
+        </common-color>
+        <common-color
+            :default-color="{ transparency: 1 }"
+            :model-value="store.mapSettings.colors?.[themeKey]?.staffedAirport ? { transparency: store.mapSettings.colors?.[themeKey]?.staffedAirport ?? 1 } : reactive({ transparency: 1 })"
+            transparency-only
+            @update:modelValue="setUserMapSettings({ colors: { [themeKey]: { staffedAirport: $event.transparency } } })"
+        >
+            Staffed Airport
+        </common-color>
+        <common-color
+            :default-color="{ transparency: 1 }"
+            :model-value="store.mapSettings.colors?.[themeKey]?.defaultAirport ? { transparency: store.mapSettings.colors?.[themeKey]?.defaultAirport ?? 1 } : reactive({ transparency: 1 })"
+            transparency-only
+            @update:modelValue="setUserMapSettings({ colors: { [themeKey]: { defaultAirport: $event.transparency } } })"
+        >
+            Unstaffed Airport
         </common-color>
 
         <common-block-title>
@@ -65,6 +105,53 @@
         >
             {{ title }}
         </common-color>
+        <common-popup
+            v-model="themeSyncActive"
+            width="100vw"
+        >
+            <template #title>
+                Settings overwrite
+            </template>
+
+            Are you sure you want to overwrite {{ isLightTheme ? 'dark' : 'light' }} theme settings?<br> This change is permanent.
+
+            <template #actions>
+                <common-button
+                    type="secondary-flat"
+                    @click="syncThemes"
+                >
+                    Confirm overwrite
+                </common-button>
+                <common-button
+                    type="secondary-875"
+                    @click="backupMapSettings()"
+                >
+                    Backup data
+                </common-button>
+                <common-button
+                    type="primary"
+                    @click="themeSyncActive = false"
+                >
+                    Cancel that
+                </common-button>
+            </template>
+        </common-popup>
+        <common-popup
+            v-model="themeSyncComplete"
+            width="300px"
+        >
+            <template #title>
+                Success
+            </template>
+            Themes have been synced.
+            <template
+                #actions
+            >
+                <common-button  @click="themeSyncComplete = false">
+                    Roger
+                </common-button>
+            </template>
+        </common-popup>
     </div>
 </template>
 
@@ -76,6 +163,7 @@ import { aircraftSvgColors } from '~/composables/pilots';
 import type { MapAircraftStatus } from '~/composables/pilots';
 import type { PartialRecord } from '~/types';
 import CommonBlockTitle from '~/components/common/blocks/CommonBlockTitle.vue';
+import { backupMapSettings } from '~/composables/settings';
 
 const store = useStore();
 
@@ -88,9 +176,24 @@ const aircraftOptions: PartialRecord<MapAircraftStatus, string> = {
     active: 'Active',
     green: 'Own aircraft',
     hover: 'Hover',
-    landed: 'Landed (dahsboard)',
+    landed: 'Landed (dashboard)',
     arriving: 'Arriving (dashboard)',
     departing: 'Departing (dashboard)',
+};
+
+const themeSyncActive = ref(false);
+const themeSyncComplete = ref(false);
+const isLightTheme = computed(() => store.getCurrentTheme === 'light');
+
+const syncThemes = () => {
+    setUserMapSettings({
+        colors: {
+            [isLightTheme.value ? 'default' : 'light']: store.mapSettings.colors![isLightTheme.value ? 'light' : 'default'],
+        },
+    });
+
+    themeSyncActive.value = false;
+    themeSyncComplete.value = true;
 };
 </script>